The browse action now support 'descendants' as a CGI-argument, which if set to '1' will produce the whole hierarchy for a given classifier point.  Useful in AJAX requests (such as producing a tag-cloud of an entire classifier).
One unknown undesirable side-effect of this at the moment is that once the value has been given, it is then remembered in the users jsession/cookie and consequently causes all subsequent requests to be fully expanded.  The tag-cloud example -- as a work around for now -- makes a second call over AJAX with descendants=0 to effectively clear the value.

git-svn-id: http://svn.greenstone.org/main/trunk/greenstone3@26508 a4f36ff5-3332-0410-bbfd-b1ff9fae65ff